Refactored unittest and introduced TestData simulator
* Clean up the test code and unify TestData into test fixtures package
* Added TestData simulator as Fluent API impls. So that caller can chain
  the method calls to the level of simulation and condition needs to the
  scenario.
