+++ app/workflow_manager/tests/factories.py
@@ -1,12 +1,12 @@
-from enum import Enum
 import uuid
-from datetime import datetime, timedelta
+from datetime import datetime
+from enum import Enum
 from zoneinfo import ZoneInfo
 
 import factory
 from django.utils.timezone import make_aware
 
-from workflow_manager.models import Workflow, WorkflowRun, Payload, Library, State, LibraryAssociation, AnalysisRun
+from workflow_manager.models import Workflow, WorkflowRun, Payload, Library, State, AnalysisRun
 
 
 class TestConstant(Enum):
@@ -77,71 +77,3 @@ class StateFactory(factory.django.DjangoModelFactory):
 class AnalysisRunFactory(factory.django.DjangoModelFactory):
     class Meta:
         model = AnalysisRun
-
-class PrimaryTestData():
-    WORKFLOW_NAME = "TestWorkflow"
-
-    STATUS_DRAFT = "DRAFT"
-    STATUS_START = "READY"
-    STATUS_RUNNING = "RUNNING"
-    STATUS_END = "SUCCEEDED"
-    STATUS_FAIL = "FAILED"
-    STATUS_RESOLVED = "RESOLVED"
-
-
-
-    def create_primary(self, generic_payload, libraries):
-        """
-        Case: a primary workflow with two executions linked to 4 libraries
-        The first execution failed and led to a repetition that succeeded
-        """
-
-        wf = WorkflowFactory(workflow_name=self.WORKFLOW_NAME + "Primary")
-
-        # The first execution (workflow run 1)
-        wfr_1: WorkflowRun = WorkflowRunFactory(
-            workflow_run_name=self.WORKFLOW_NAME + "PrimaryRun1",
-            portal_run_id="1234",
-            workflow=wf
-        )
-
-        for i, state in enumerate([self.STATUS_DRAFT, self.STATUS_START, self.STATUS_RUNNING, self.STATUS_FAIL]):
-            StateFactory(workflow_run=wfr_1, status=state, payload=generic_payload,
-                         timestamp=make_aware(datetime.now() + timedelta(hours=i)))
-        for i in [0, 1, 2, 3]:
-            LibraryAssociation.objects.create(
-                workflow_run=wfr_1,
-                library=libraries[i],
-                association_date=make_aware(datetime.now()),
-                status="ACTIVE",
-            )
-
-        # The second execution (workflow run 2)
-        wfr_2: WorkflowRun = WorkflowRunFactory(
-            workflow_run_name=self.WORKFLOW_NAME + "PrimaryRun2",
-            portal_run_id="1235",
-            workflow=wf
-        )
-        for i, state in enumerate([self.STATUS_DRAFT, self.STATUS_START, self.STATUS_RUNNING, self.STATUS_END]):
-            StateFactory(workflow_run=wfr_2, status=state, payload=generic_payload,
-                         timestamp=make_aware(datetime.now() + timedelta(hours=i)))
-        for i in [0, 1, 2, 3]:
-            LibraryAssociation.objects.create(
-                workflow_run=wfr_2,
-                library=libraries[i],
-                association_date=make_aware(datetime.now()),
-                status="ACTIVE",
-            )
-
-    def setup(self):
-
-        # Common components: payload and libraries
-        generic_payload = PayloadFactory()  # Payload content is not important for now
-        libraries = [
-            LibraryFactory(orcabus_id="01J5M2JFE1JPYV62RYQEG99CP1", library_id="L000001"),
-            LibraryFactory(orcabus_id="02J5M2JFE1JPYV62RYQEG99CP2", library_id="L000002"),
-            LibraryFactory(orcabus_id="03J5M2JFE1JPYV62RYQEG99CP3", library_id="L000003"),
-            LibraryFactory(orcabus_id="04J5M2JFE1JPYV62RYQEG99CP4", library_id="L000004")
-        ]
-
-        self.create_primary(generic_payload, libraries)
